ds-sensors-mqtt/requirements.txt
paho-mqtt==1.5.1
pytest==6.2.4
python-dotenv==0.19.0
w1thermsensor==2.0.0
